Vidos Entegrasyon 19.12.2025/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-Mail==0.10.0
itsdangerous==2.2.0
requests==2.32.3
xmltodict==0.13.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
openpyxl==3.1.5
scikit-learn==1.5.1
rapidfuzz==3.9.4
APScheduler==3.10.4
pandas==2.2.0
Flask-Migrate==4.0.7
Pillow==10.2.0
gunicorn==22.0.0
psycopg2-binary==2.9.9